Add support for floating point numbers

+++ calc.py
@@ -6,8 +6,8 @@ def compute(input_string):
 
 
     values = input_string.split(' ')
-    num0 = int(values[0])
-    num1 = int(values[2])
+    num0 = float(values[0])
+    num1 = float(values[2])
     operator = values[1]
     if operator == '+':
         return num0 + num1
